requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-static==0.0.6
Django==3.0.8
django-taggit==1.3.0
djangorestframework==3.11.0
docopt==0.6.2
gunicorn==19.1.1
idna==2.10
mysqlclient==2.0.1
pipreqs==0.4.10
psycopg2-binary==2.8.5
PyMySQL==0.9.3
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.9
yarg==0.1.9
